# internal/models/sql_models/subscription.go
package sql_models

import "time"

// sql_models.Subscription model
// @Description Subscription information
type Subscription struct {
	ID          string     `db:"id"`
	ServiceName string     `db:"service_name"`
	Price       int        `db:"price"`
	UserID      string     `db:"user_id"`
	StartDate   time.Time  `db:"start_date"`
	EndDate     *time.Time `db:"end_date"`
	CreatedAt   time.Time  `db:"created_at"`
}